backend/requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
Authlib==1.2.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
Deprecated==1.2.14
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
environs==9.5.0
fastapi==0.103.1
fastapi_cors==0.0.6
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
limits==3.6.0
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
orjson==3.9.7
packaging==23.2
passlib==1.7.4
psycopg2==2.9.7
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.1
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.1
rsa==4.9
six==1.16.0
slowapi==0.1.8
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
wrapt==1.15.0
